+++ src/main/java/com/dbtechschool/repository/CourseRepository.java
@@ -11,10 +11,10 @@ import org.springframework.transaction.annotation.Transactional;
 import com.dbtechschool.model.Course;
 
 @Repository
-public interface CourseRepository extends CrudRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long> {
 
 	@Modifying
 	@Transactional
 	@Query("update Course c set c.name = ?1, c.start = ?2, c.end = ?3, c.tags = ?4, c.description = ?5, c.difficulty = ?6  where c.id = ?7")
-	int updateCourseById(String name, Date start, Date end, String tags, String description, char difficulty, int id);
+	int updateCourseById(String name, Date start, Date end, String tags, String description, char difficulty, Long id);
 }
